requirements.txt
asgiref==3.5.2
autopep8==1.6.0
black==22.6.0
click==8.1.3
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.9.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
zipp==3.8.1
